refactor(modules/auth): move save user login to `authService`
