+++ src/modules/auth/auth.service.ts
@@ -1,89 +1,48 @@
-import { HttpService } from '@nestjs/axios'
 import { Injectable } from '@nestjs/common'
-import { ConfigService } from '@nestjs/config'
-import { JwtService } from '@nestjs/jwt'
-import { Profile as PassportSpotifyProfile } from 'passport-spotify'
-import { map, catchError, firstValueFrom } from 'rxjs'
+import { AccessToken } from '@spotify/web-api-ts-sdk'
 
-import { SecretData } from './dtos'
-import { TokenOptions } from './types'
+import { AuthorizeParams } from './types'
 
-import { Environment } from '@config/environment'
-import { SpotifyToken, Profile, SpotifyProfile } from '@common/types/spotify'
-import { applyAuthorizationHeader, catchSpotifyError } from '@common/utils'
-import { adaptProfile, adaptSecretData } from '@common/adapters'
+import { UsersRepository } from '@modules/users'
+import { ProfilesService } from '@modules/profiles'
+import { SpotifyUsersService } from '@modules/spotify/users'
 
 @Injectable()
 export class AuthService {
   constructor(
-    private readonly jwtService: JwtService,
-    private readonly httpService: HttpService,
-    private readonly configService: ConfigService
+    private readonly usersRepository: UsersRepository,
+    private readonly profilesService: ProfilesService,
+    private readonly spotifyUsersService: SpotifyUsersService
   ) {}
 
-  login({ id, username }: PassportSpotifyProfile) {
-    const payload = {
-      name: username,
-      sub: id,
+  async saveUser(token: AccessToken): Promise<AuthorizeParams> {
+    const spotifyProfile = await this.spotifyUsersService.profile(token)
+
+    const foundUser = await this.usersRepository.findOneByProfileId(
+      spotifyProfile.id
+    )
+
+    const { access_token: accessToken, refresh_token: refreshToken } = token
+
+    let id: string
+
+    if (foundUser) {
+      id = foundUser.id
+    } else {
+      const profile = await this.profilesService.create(spotifyProfile)
+
+      const createdUser = await this.usersRepository.createUser({
+        profile,
+        refreshToken,
+      })
+
+      id = createdUser.id
     }
 
-    return this.jwtService.sign(payload)
-  }
-
-  token({ refreshToken, code }: TokenOptions): Promise<SecretData> {
-    const url = `${this.configService.get(
-      Environment.SPOTIFY_ACCOUNTS_URL
-    )}/api/token`
-    const cliendId = this.configService.get<string>(
-      Environment.SPOTIFY_CLIENT_ID
-    )
-    const clientSecret = this.configService.get<string>(
-      Environment.SPOTIFY_CLIENT_SECRET
-    )
-    const callbackUrl = this.configService.get<string>(
-      Environment.SPOTIFY_CALLBACK_URL
-    )
-
-    const bufferedCredentials = Buffer.from(
-      `${cliendId}:${clientSecret}`
-    ).toString('base64')
-    const params = new URLSearchParams()
-
-    if (refreshToken) {
-      params.append('refresh_token', refreshToken)
-      params.append('grant_type', 'refresh_token')
+    return {
+      accessToken,
+      refreshToken,
+      id,
     }
-    if (code) {
-      params.append('code', code)
-      params.append('grant_type', 'authorization_code')
-      callbackUrl && params.append('redirect_uri', callbackUrl)
-    }
-
-    return firstValueFrom(
-      this.httpService
-        .post<SpotifyToken>(url, params, {
-          headers: {
-            'Content-Type': 'application/x-www-form-urlencoded',
-            Authorization: `Basic ${bufferedCredentials}`,
-          },
-        })
-        .pipe(
-          map(response => response.data),
-          map(adaptSecretData),
-          catchError(catchSpotifyError)
-        )
-    )
-  }
-
-  profile(accessToken: string): Promise<Profile> {
-    return firstValueFrom(
-      this.httpService
-        .get<SpotifyProfile>('/me', applyAuthorizationHeader(accessToken))
-        .pipe(
-          map(response => response.data),
-          map(adaptProfile),
-          catchError(catchSpotifyError)
-        )
-    )
   }
 }
