+++ src/modules/auth/types/authorize-params.type.ts
@@ -1,0 +1,5 @@
+export interface AuthorizeParams {
+  refreshToken: string
+  accessToken: string
+  id: string
+}
