+++ src/modules/auth/types/index.ts
@@ -1,3 +1,2 @@
-export * from './jwt-payload'
-export * from './redirect-response'
-export * from './token-options'
+export * from './authorize-params.type'
+export * from './redirect-response.type'
